added functionality to make a todo done
