+++ index.html
@@ -10,10 +10,10 @@
     <div class="container">
       <h1>Keep What Todo</h1>
       <input type="text" id="taskInput" placeholder="Type task" />
-      <button id="addTask" class="addButton" type="submit">Add New</button>
+      <button id="addTask" class="addButtonStyle" type="submit">Add New</button>
       <p id="errorMessage"></p>
       <div class="list-container">
-        <h4 class="list-title">Todos</h4>
+        <h4 id="listTitle" class="hide">Todos</h4>
         <ul id="taskList"></ul>
       </div>
     </div>
