+++ index.js
@@ -1,4 +1,9 @@
-import { addTaskButton, taskInputText, taskList } from "./scripts/elements.js";
+import {
+  addTaskButton,
+  listTitle,
+  taskInputText,
+  taskList,
+} from "./scripts/elements.js";
 import { handleCreateTodo } from "./scripts/addTask.js";
 import { todos } from "./scripts/deteleTask.js";
 import { createTaskElement } from "./scripts/taskActions.js";
@@ -11,6 +16,13 @@ export const renderTodoList = () => {
     const newTask = createTaskElement(todo);
     taskList.appendChild(newTask);
   });
+  if (todos.length > 0) {
+    listTitle.classList.remove("hide");
+    listTitle.classList.add("show");
+    return;
+  }
+  listTitle.classList.remove("show");
+  listTitle.classList.add("hide");
 };
 
 addTaskButton.addEventListener("click", handleCreateTodo);
