+++ scripts/doneTask.js
@@ -1,0 +1,8 @@
+import { renderTodoList } from "../index.js";
+import { todos } from "./deteleTask.js";
+
+export const handleDone = (taskId) => {
+  const task = todos.find((todo) => todo.id === taskId);
+  task.isDone = true;
+  renderTodoList();
+};
