+++ scripts/elements.js
@@ -2,3 +2,4 @@ export const addTaskButton = document.getElementById("addTask");
 export const taskInputText = document.getElementById("taskInput");
 export const taskList = document.getElementById("taskList");
 export const errorMessage = document.getElementById("errorMessage");
+export const listTitle = document.getElementById("listTitle");
